scripts/requirements.txt
pandas
SQLAlchemy>=2.0
pyodbc
python-dotenv
tenacity
azure-storage-blob>=12.20
